Set up default view and expanded view in Parent and moved states to parent
